Allow downloading images
If the user passes a url instead of a local path, it will be downloaded
and then it will be flashed.

This is useful since generally these files need to come from some url it
saves the download trip and it allows the app to manage that file that
is rather useless afterwards.